refactor: ui.model直下にあったViewModelすべてにそれぞれの階層を追加
