+++ app/src/main/java/com/punyo/nitechvacancyviewer/ui/buildingvacancy/VacancyComponentViewModel.kt
@@ -1,4 +1,4 @@
-package com.punyo.nitechvacancyviewer.ui.model
+package com.punyo.nitechvacancyviewer.ui.buildingvacancy
 
 import androidx.lifecycle.ViewModel
 import com.punyo.nitechvacancyviewer.data.building.BuildingRepository
