+++ app/src/main/java/com/punyo/nitechvacancyviewer/ui/initialize/InitializeScreen.kt
@@ -1,4 +1,4 @@
-package com.punyo.nitechvacancyviewer.ui.screen
+package com.punyo.nitechvacancyviewer.ui.initialize
 
 import android.app.Activity
 import androidx.annotation.StringRes
@@ -34,7 +34,6 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.punyo.nitechvacancyviewer.R
 import com.punyo.nitechvacancyviewer.ui.component.CampusSquareWebViewComponent
 import com.punyo.nitechvacancyviewer.ui.component.LoadingProgressIndicatorComponent
-import com.punyo.nitechvacancyviewer.ui.model.InitializeScreenViewModel
 import com.punyo.nitechvacancyviewer.ui.theme.AppTheme
 
 @Composable
