+++ app/src/main/java/com/punyo/nitechvacancyviewer/ui/roomvacancy/RoomVacancyScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.punyo.nitechvacancyviewer.ui.model
+package com.punyo.nitechvacancyviewer.ui.roomvacancy
 
 import androidx.lifecycle.ViewModel
 import com.punyo.nitechvacancyviewer.application.enums.RoomVacancyStatus
