+++ app/src/main/java/com/punyo/nitechvacancyviewer/ui/signin/SignInScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.punyo.nitechvacancyviewer.ui.model
+package com.punyo.nitechvacancyviewer.ui.signin
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
